POC-APP/requirements.txt
boto3
pandas
openpyxl
streamlit
Pillow
python-dotenv
xlsxwriter
google-api-python-client
anthropic
textblob
scikit-learn
